V2/requirements.txt
<<<<<<< HEAD
aiohttp==3.9.1
asyncio-mqtt==0.16.1
ccxt==4.3.56
numpy==1.24.3
pandas==2.0.3
python-socketio==5.10.0
requests==2.31.0
scikit-learn==1.3.0
joblib==1.3.2
websockets==12.0
python-dotenv==1.0.0
uvloop==0.19.0

=======
python-socketio[asyncio_client]
aiohttp>=3.0.0
scikit-learn>=0.24 # Incluye joblib. Ajustar versión si es necesario.
ccxt>=2.0.0 # Usar una versión reciente de ccxt para async support
# numpy>=1.19 # Si se usa numpy explícitamente
pandas>=1.1.0 # Si se usa pandas explícitamente
# tensorflow # Si se usa TensorFlow/Keras
#socketio>=0.2.1 # Asegurarse de que sea compatible con python-socketio
>>>>>>> jules/multi-fixes-optimizations
